react-ify'd the code in topicPicker
made of 8 commits:
separate collapsibleList out into its own react element
generate topics from topic json
rename files to match nice conventions

TitleCase for components and their assets
move annotation view components into components/annoation
squash ArticleHeader into Article
use proper new filepaths
remove console.log
