+++ app/styles/topicpicker.scss
@@ -1,12 +1,6 @@
 @import 'styleguide';
 @import 'common';
 
-@mixin ul-reset {
-  list-style: none;
-  padding-left: initial;
-  margin-bottom: initial;
-}
-
 $topic-picker-nav-width: 25%;
 
 .topic-picker {
@@ -149,48 +143,5 @@ $topic-picker-nav-width: 25%;
   &__topic {
     padding: 5px 10px;
     line-height: 1.5;
-  }
-}
-
-.variable-list {
-  background: #fff;
-  padding: 5px 10px;
-  margin-top: 15px;
-
-  ul {
-    @include ul-reset;
-    padding-left: 13px;
-  }
-
-  &__collapsible {
-    &:before {
-      display: inline-block;
-      width: 8px;
-      height: 8px;
-      margin-right: 5px;
-      background:
-        linear-gradient(to bottom,
-                        transparent 35%,
-                        $offblack 35%,
-                        $offblack 65%,
-                        transparent 65%);
-      content: '';
-    }
-
-    &--closed {
-      &:before {
-        background:
-        linear-gradient(to bottom,
-                        transparent 35%,
-                        $offblack 35%,
-                        $offblack 65%,
-                        transparent 65%),
-        linear-gradient(to right,
-                        transparent 35%,
-                        $offblack 35%,
-                        $offblack 65%,
-                        transparent 65%);
-      }
-    }
   }
 }
